Subscriber buffer queue bug fix
Fixed a bug that split messages and confused PackerProtocol when the bufer
queue was full. We cannot skip the buffers added to the queue (using
TryAdd) since it may be part of a larger message.

Additionaly BufferPool is made instance making sure if it went out of
scope all the bufferes held were cleared.